html/browsers/history/the-location-interface/location_pathname.html

<!DOCTYPE HTML>
<html>
  <head>
    <title>location_pathname</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
    if (('location' in window) && window.location != null
      && ('location' in document) && document.location != null){
      var pathname = location.pathname;
      var url = location.href

      url = url.replace(location.protocol + "//" + location.host, "");

      test(function() {
        assert_equals(pathname, url, "pathname attribute is not correct.");
      });
    } else {
      test(function() {
        assert_unreached("location interface is not supported.");
      });
    }
    </script>
  </body>
</html>
